+++ test/e2e/specs/editor/blocks/navigation.spec.js
@@ -7,15 +7,19 @@ test.describe(
 	'As a user I want the navigation block to fallback to the best possible default',
 	() => {
 		test.beforeAll( async ( { requestUtils } ) => {
-			await requestUtils.activateTheme( 'emptytheme' );
+			//TT3 is preferable to emptytheme because it already has the navigation block on its templates.
+			await requestUtils.activateTheme( 'twentytwentythree' );
 		} );
 
-		test.beforeEach( async ( { admin } ) => {
-			await admin.createNewPost();
+		test.beforeEach( async ( { requestUtils } ) => {
+			await Promise.all( [ requestUtils.deleteAllMenus() ] );
 		} );
 
 		test.afterAll( async ( { requestUtils } ) => {
-			await requestUtils.activateTheme( 'twentytwentyone' );
+			await Promise.all( [
+				requestUtils.deleteAllMenus(),
+				requestUtils.activateTheme( 'twentytwentyone' ),
+			] );
 		} );
 
 		test.afterEach( async ( { requestUtils } ) => {
@@ -23,8 +27,10 @@ test.describe(
 		} );
 
 		test( 'default to a list of pages if there are no menus', async ( {
+			admin,
 			editor,
 		} ) => {
+			await admin.createNewPost();
 			await editor.insertBlock( { name: 'core/navigation' } );
 
 			const pageListBlock = editor.canvas.getByRole( 'document', {
@@ -48,5 +54,121 @@ test.describe(
 <!-- /wp:navigation -->`
 			);
 		} );
+
+		test( 'default to my only existing menu', async ( {
+			admin,
+			editor,
+			page,
+			requestUtils,
+		} ) => {
+			await admin.createNewPost();
+			const createdMenu = await requestUtils.createNavigationMenu( {
+				title: 'Test Menu 1',
+				content:
+					'<!-- wp:navigation-link {"label":"WordPress","type":"custom","url":"http://www.wordpress.org/","kind":"custom","isTopLevelLink":true} /-->',
+			} );
+
+			await editor.insertBlock( { name: 'core/navigation' } );
+
+			// Check the block in the canvas.
+			await expect(
+				editor.canvas.locator(
+					`role=textbox[name="Navigation link text"i] >> text="WordPress"`
+				)
+			).toBeVisible();
+
+			// Check the markup of the block is correct.
+			await editor.publishPost();
+			await expect.poll( editor.getBlocks ).toMatchObject( [
+				{
+					name: 'core/navigation',
+					attributes: { ref: createdMenu.id },
+				},
+			] );
+			await page.locator( 'role=button[name="Close panel"i]' ).click();
+
+			// Check the block in the frontend.
+			await page.goto( '/' );
+			await expect(
+				page.locator(
+					`role=navigation >> role=link[name="WordPress"i]`
+				)
+			).toBeVisible();
+		} );
+
+		test( 'default to the only existing classic menu if there are no block menus', async ( {
+			admin,
+			page,
+			editor,
+			requestUtils,
+		} ) => {
+			// Create a classic menu.
+			await requestUtils.createClassicMenu( 'Test Classic 1' );
+			await admin.createNewPost();
+
+			await editor.insertBlock( { name: 'core/navigation' } );
+			// We need to check the canvas after inserting the navigation block to be able to target the block.
+			await expect.poll( editor.getBlocks ).toMatchObject( [
+				{
+					name: 'core/navigation',
+				},
+			] );
+
+			// Check the block in the canvas.
+			await expect(
+				editor.canvas.locator(
+					`role=textbox[name="Navigation link text"i] >> text="Custom link"`
+				)
+			).toBeVisible();
+
+			// Check the block in the frontend.
+			await page.goto( '/' );
+
+			await expect(
+				page.locator(
+					`role=navigation >> role=link[name="Custom link"i]`
+				)
+			).toBeVisible();
+		} );
 	}
 );
+
+test.describe( 'Navigation block', () => {
+	test.describe(
+		'As a user I want to see a warning if the menu referenced by a navigation block is not available',
+		() => {
+			test.beforeEach( async ( { admin } ) => {
+				await admin.createNewPost();
+			} );
+
+			test( 'warning message shows when given an unknown ref', async ( {
+				editor,
+			} ) => {
+				await editor.insertBlock( {
+					name: 'core/navigation',
+					attributes: {
+						ref: 1,
+					},
+				} );
+
+				// Check the markup of the block is correct.
+				await editor.publishPost();
+
+				await expect.poll( editor.getBlocks ).toMatchObject( [
+					{
+						name: 'core/navigation',
+						attributes: { ref: 1 },
+					},
+				] );
+
+				// Find the warning message
+				const warningMessage = editor.canvas
+					.getByRole( 'document', { name: 'Block: Navigation' } )
+					.getByText(
+						'Navigation menu has been deleted or is unavailable.'
+					);
+				await expect( warningMessage ).toBeVisible();
+			} );
+		}
+	);
+} );